moved component from ui folder to component folder and fixed import paths
